+++ packages/components/CHANGELOG.md
@@ -25,12 +25,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 - `Link Tile`, `Button Tile`: added `footer` slot
   ([#4025](https://github.com/porsche-design-system/porsche-design-system/pull/4025))
 - `Radio Group`: ([#4042](https://github.com/porsche-design-system/porsche-design-system/pull/4042))
+- `Radio Group`: `blur` event ([#4057](https://github.com/porsche-design-system/porsche-design-system/pull/4057))
+- `Flyout`: Prop `backdrop` ([#4058](https://github.com/porsche-design-system/porsche-design-system/pull/4058))
 
 ### Changed
 
 - `Radio Button Wrapper`: component is deprecated. Use `Radio Group` instead.
   ([#4042](https://github.com/porsche-design-system/porsche-design-system/pull/4042))
-
 
 ### Fixed